// frontend/src/pages/renters/interfaces/renter-errors.interface.ts
export interface RenterErrors {
  name?: string;
  nit?: string;
  address?: string;
  city?: string;
  email?: string;
  phone?: string;
  password?: string;
  legalRepresentative?: string;
  planId?: string;
  planExpiresAt?: string;
  balance?: string;
  lowBalanceThreshold?: string;
  lowBalanceAlertEnabled?: string;
  status?: string;
}